% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_results.R
\name{combine_results}
\alias{combine_results}
\title{Results preprocessing}
\usage{
combine_results(
  fileNamesPattern,
  path = "./",
  saveToFile = NA_character_,
  mergeResultsWithConditions = FALSE,
  keepFileNames = FALSE,
  trimIterationsNumber = FALSE
)
}
\arguments{
\item{fileNamesPattern}{a \link{regex}(regular expression) matching names of the
files that should be combined}

\item{path}{optionally a path to the directory in which these files are
stored (if it is different than the current working directory)}

\item{saveToFile}{optionally the name of the (.RData) file to save the
combined results to}

\item{mergeResultsWithConditions}{a logical value indicating whether
variables describing specific values of the data-generating model parameters
in each simulation condition should be appended to objects storing simulation
results (if not, only the simulation condition number will be stored in these
objects, allowing such a merge to be performed in the future)}

\item{keepFileNames}{a logical value indicating whether information about
the source file should be stored with the combined results}

\item{trimIterationsNumber}{a logical value indicating whether to trim the
number of iterations within each simulation condition to the smallest number
of iterations among all conditions in the results}
}
\value{
a list of combined objects
}
\description{
Enables to combine simulation results stored in different files.
}
\details{
Please note that the function assigns new (ordinal) numbers to the simulation
conditions and iterations to make them unique among all combined files.
To emphasize this fact the column storing ordinal number of condition
(condition id) is named \emph{condition} in the combined results (instead of
\emph{cond}) and the column storing ordinal number of interation (iteration id)
is named \emph{iter} in the combined results (instead of \emph{i}).
}
