% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{compute_dep_std_var}
\alias{compute_dep_std_var}
\title{Data generation}
\usage{
compute_dep_std_var(nLV, nLI, r, slope, slopeI, errorStdDev)
}
\arguments{
\item{nLV}{number of exogenous latent variables}

\item{nLI}{number of latent interactions}

\item{r}{correlation between (each pair of) exogenous latent variables}

\item{slope}{a scalar - value of the slope coefficient for the main effects
(typically set to 1)}

\item{slopeI}{a scalar - value of the slope coefficient for the interaction
effects, typically returned by the \link{compute_interaction_slope}}

\item{errorStdDev}{a scalar - standard deviation of the dependent latent
variable error term, typically returned by the \link{compute_error_std_dev}}
}
\value{
a scalar
}
\description{
Computes standard deviation of the latent dependent variable so it can be
standardized to have variance of 1 before returning it by the \link{generate_latent}
}
