% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_latent.R
\name{compute_error_std_dev}
\alias{compute_error_std_dev}
\title{Data generation}
\usage{
compute_error_std_dev(nLV, r, partR, slope = 1)
}
\arguments{
\item{nLV}{number of exogenous latent variables}

\item{r}{correlation between (each pair of) exogenous latent variables}

\item{partR}{partial correlation between (each) exogenous latent variable and
dependent latent variable}

\item{slope}{a scalar - value of the slope coefficient for the main effects
(typically set to 1)}
}
\value{
a scalar
}
\description{
Computes the (non-normalized, compare \link{compute_dep_std_var}) standard
deviation of the latent dependent variable's error term (used internally by
the \link{generate_latent})
}
