% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{compute_interaction_slope}
\alias{compute_interaction_slope}
\title{Data generation}
\usage{
compute_interaction_slope(nLI, r, partRI, errorStdDev)
}
\arguments{
\item{nLI}{number of latent interactions}

\item{r}{correlation between (each pair of) exogenous latent variables}

\item{partRI}{partial correlation between (each) latent interaction variable
and dependent latent variable}

\item{errorStdDev}{a scalar - standard deviation of the dependent latent
variable error term, typically returned by the \link{compute_error_std_dev}}
}
\value{
a scalar
}
\description{
Computes the value of the slope coefficient for the interaction latent
variables that should be used while generating the dependent latent variable
(used internally by the \link{generate_latent})
}
