% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conditions}
\alias{conditions}
\title{Simulation conditions used in our study}
\format{
A data frame with 144 rows and 8 columns.
See \link{check_conditions} for the columns description.
}
\usage{
conditions
}
\description{
Simulation conditions used in our study
}
\keyword{datasets}
