% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_interaction_indicators.R
\name{create_interaction_indicators}
\alias{create_interaction_indicators}
\title{Data transformation}
\usage{
create_interaction_indicators(
  data,
  indicatorsMapping,
  interactionsMapping,
  center = c("mean", "none", "residual"),
  nonredundantOnly = TRUE
)
}
\arguments{
\item{data}{a matrix with values of observed indicators, typically generated
by the \link{generate_observed}}

\item{indicatorsMapping}{a numeric (or logical) matrix with indicator names
in rows and latent variable names in columns describing which indicators are
loaded by which latent variables (non-zero cells indicate the relationship);
typically a matrix of factor loadings generated by the \link{generate_loadings}}

\item{interactionsMapping}{a numeric (or logical) matrix with latent
interaction names in rows and latent variable names in columns describing
which latent variables create interactions (non-zero cells indicate the
relationship); typically the attribute \code{mapping} of a result of calling the
\link{generate_latent}}

\item{center}{a string specifying type of centering: grand mean centering
("mean", the default), no centering ("none") or residual centering ("residual")}

\item{nonredundantOnly}{a logical value specifying whether each latent
variable indicator should create only one interaction observed indicator (in
a given interaction); should be set to \code{TRUE} for the \emph{unconstrained}
approach; otherwise all the combinations of indicators are used to create
interaction's observed indicators (and consequently constraints on their
random error terms should be imposed in the model specification)}
}
\value{
\code{data} matrix with additional columns representing latent
interactions indicators; matrix with mapping of the indicators to latent
variables (including newly created latent interaction ones) is returned as
an attribute
}
\description{
Computes observed indicators to be loaded by the latent variables
representing latent interactions
}
\details{
Function can deal also with interactions of higher order, i.e. those for
which there are more than two non-zero entries in the given row of the
\code{interactionsMapping}.
}
\examples{
latent <- generate_latent(nObs = 10^5, nLV = 3, nLI = 2,
                          r = 0.3, partR = 0.3, partRI = 0.2)
loadings <- generate_loadings(c(x1 = 2, x2 = 2, x3 = 2, y = 3), 0.7)
observed <- generate_observed(latent, loadings,
                              latentVariances = diag(attributes(latent)$cov))
head(observed)
observed <- create_interaction_indicators(observed, loadings,
                                          attributes(latent)$mapping,
                                          center = "mean",
                                          nonredundantOnly = FALSE)
head(observed)
}
