% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_results.R
\name{enumerate_conditions}
\alias{enumerate_conditions}
\title{Results preprocessing}
\usage{
enumerate_conditions(conditions)
}
\arguments{
\item{conditions}{a data frame with combined simulation conditions
descriptions}
}
\value{
provided data frame with added column \emph{condition}
}
\description{
Provides unique numbering of conditions among multiple files that store
simulation results, even if they were ran using different sets of simulation
conditions.
}
