% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_results.R
\name{enumerate_iterations}
\alias{enumerate_iterations}
\title{Results preprocessing}
\usage{
enumerate_iterations(resultsList)
}
\arguments{
\item{resultsList}{a list of data frames with combined simulation results}
}
\value{
a data frame with columns \emph{condition}, \emph{i}, \emph{file} and \emph{iter}
}
\description{
Provides unique numbering of iterations among multiple files that store
simulation results.
}
