% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_latent.R
\name{generate_latent}
\alias{generate_latent}
\title{Data generation}
\usage{
generate_latent(nObs, nLV, nLI, r, partR, partRI = partR)
}
\arguments{
\item{nObs}{number of observations to be generated}

\item{nLV}{number of exogenous latent variables}

\item{nLI}{number of latent interactions}

\item{r}{correlation between (each pair of) exogenous latent variables}

\item{partR}{partial correlation between (each) exogenous latent variable and
dependent latent variable}

\item{partRI}{partial correlation between (each) latent interaction variable
and dependent latent variable}
}
\value{
a numeric matrix with \code{nObs} rows and \code{nLV + nLI + 1} columns
(values of \code{r}, \code{partR}, \code{partRI}, as well as mapping of latent variables
on interaction variables, slope parameters and error term standard deviation
used to generate the dependent latent variable, and covariance matrix of
latent variables are returned as attributes)
}
\description{
Generates a matrix with the sampled values of latent variables
}
\examples{
x <- generate_latent(nObs = 10^5, nLV = 3, nLI = 2,
                     r = 0.3, partR = 0.3, partRI = 0.2)
# a matrix to store partial correlations
res = diag(ncol(x))
colnames(res) = rownames(res) = colnames(x)
for (i in seq_len(ncol(res))) {
  for (j in seq_len(i - 1L)) {
    res[i, j] = res[j, i] = cor(resid(lm(x[, i] ~ 0 + x[, -c(i, j)])),
                                resid(lm(x[, j] ~ 0 + x[, -c(i, j)])))
  }
}
# correlations
round(cor(x), 2)
# partial correlations
round(res, 3)
# comparison of theoretical and empirical covariances
round(attributes(x)$cov, 2)
round(cov(x), 2)
}
\seealso{
\link{compute_error_std_dev}, \link{compute_interaction_slope},
\link{compute_dep_std_var}
}
