% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_observed.R
\name{generate_loadings}
\alias{generate_loadings}
\title{Data generation}
\usage{
generate_loadings(nIndic, fun, args = NULL)
}
\arguments{
\item{nIndic}{a named integer vector: values give a number of observed
indicators for a latent variable specified by a given element's name}

\item{fun}{either:
\itemize{
\item{a number specifying the value to which all loadings will be set}
\item{a function that should be used to generate loadings for all the
latent variables}
\item{a named list of functions to be used to generate loadings for
a specific latent variables (names of the list elements must match
names of the \code{nIndic} elements)}
}}

\item{args}{optionally either:
\itemize{
\item{a list of arguments to be passed to the \code{fun} (if \code{fun} is
a function)}
\item{a named list of lists of arguments to be used to functions given by
the \code{fun} argument (names of the list elements must match
names of the \code{fun} elements)}
}}
}
\value{
a numeric matrix of loadings with \code{sum(nIndic)} rows
and \code{length(nIndic)} columns
}
\description{
Generates a matrix of factor loadings (do \strong{not} allow cross-loadings)
}
\details{
Please note that function does \strong{not} perform any checks on the generated
values of loadings (specifically, whether they are valid standardized
loadings).
}
\examples{
# The same value of 0.7 for all the latent variables and indicators:
generate_loadings(c(x1 = 3, x2 = 4, y = 5), 0.7)
# Different values for different latent variables, but constant across indicators:
generate_loadings(c(x1 = 3, x2 = 4, y = 5), list(x1 = I, x2= I, y = I),
                  list(x1 = list(0.7), x2 = list(0.5), y = list(0.6)))
# Loadings sampled from the uniform distribution:
generate_loadings(c(x1 = 3, x2 = 4, y = 5),
                  list(x1 = runif, x2= runif, y = runif),
                  list(x1 = list(n = 3, min = 0.5, max = 0.7),
                       x2 = list(n = 4, min = 0.5, max = 0.7),
                       y = list(n = 5, min = 0.5, max = 0.7)))
}
\seealso{
\link{generate_observed}, \link{generate_latent}
}
