% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_observed.R
\name{generate_observed}
\alias{generate_observed}
\title{Data generation}
\usage{
generate_observed(
  latent,
  loadings,
  intercepts = rep(0, nrow(loadings)),
  loadingsStandardized = TRUE,
  latentVariances = rep(1, ncol(latent))
)
}
\arguments{
\item{latent}{a matrix with values of latent variables, typically generated
by the \link{generate_latent}}

\item{loadings}{a matrix with values of factor loadings (with observed
indicators in rows and latent variables in columns), typically generated by
the \link{generate_loadings}}

\item{intercepts}{optionally a numeric vector of the observed indicators
intercept parameters (of the same length as the number of rows in \code{loadings});
by default, the intercept of 0 is assumed for all indicators}

\item{loadingsStandardized}{a logical value indicating whether factor
loadings given by \code{loadings} are standardized ones (\code{TRUE} by default);
\strong{if \code{FALSE}, variance of the error terms of observed indicators will be
fixed to 1}}

\item{latentVariances}{optionally a vector of variances of the latent
variables - \strong{needed only if} \code{loadingsStandardized = TRUE};
by default it is assumed that latent variables have variance of 1}
}
\value{
a numeric matrix of observed indicators' values with \code{nrow(latent)}
rows and \code{nrow(loadings)} columns; also, a matrix describing which observed
indicator is loaded by which latent variable(s) and matrix of unstandardized
loadings are returned as attributes
}
\description{
Generates a matrix of values of the latent variables' observed indicators
}
\details{
Cross-loadings are not supported with \code{loadingsStandardized=TRUE}.
}
\examples{
latent <- generate_latent(nObs = 10^5, nLV = 3, nLI = 2,
                          r = 0.3, partR = 0.3, partRI = 0.2)
loadings <- generate_loadings(c(x1 = 3, x2 = 3, x3 = 3, y = 4), 0.7)
observed <- generate_observed(latent, loadings)
head(observed)
}
