% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_something.R
\name{get_model_pars}
\alias{get_model_pars}
\title{Extracting results from the estimated models}
\usage{
get_model_pars(x)
}
\arguments{
\item{x}{an object of one of the following classes: \emph{lavaan}, \emph{mplusObject},
\emph{miive} or \emph{seminr_model}}
}
\value{
a list of two data frames:

\strong{structural}
\describe{
\item{DV}{dependent latent variable}
\item{IV}{independent latent variable (predictor)}
\item{est}{regression (slope) coefficient}
\item{se}{standard error of \code{est}}
\item{std.all}{standardized regression (slope) coefficient}
}
\strong{measurement}
\describe{
\item{LV}{latent variable}
\item{obsIndic}{observed indicator}
\item{est}{factor loading}
\item{se}{standard error of \code{est}}
\item{lambda}{standardized factor loading}
}
}
\description{
Gets parameters from estimated models.
}
\details{
\itemize{
\item{For \emph{seminr_model} objects (PLS models) only standardized structural
coefficients are reported in model results so in \code{structural} values
in \code{est} always equal those in \code{std.all} for these models.}
\item{For \emph{miive} objects no loadings are reported for first observed
indicators of each latent variable as these serve as so-called
\emph{scaling indicators} and play a special role in model parameters
estimation.}
}
}
