% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_something.R
\name{get_model_summary}
\alias{get_model_summary}
\title{Extracting results from the estimated models}
\usage{
get_model_summary(x)
}
\arguments{
\item{x}{an object of one of the following classes: \emph{lavaan}, \emph{mplusObject},
\emph{miive} or \emph{seminr_model}}
}
\value{
a data frame with the following columns:
\describe{
\item{time}{estimation time in seconds (only for \emph{lavaan} and \emph{mplusObject})}
\item{converged}{logical value indicating whether model converged}
\describe{
\item{\emph{lavaan}}{taken from information included explicitly in the object}
\item{\emph{mplusObject}}{determined by checking whether standard errors (or
posterior standard deviations) of model parameters
are reported in the results - if yes, it is assumed
that model converged}
\item{\emph{miive}}{determined by checking whether parameters covariance
matrix is reported for every equation estimated in the
model - if yes, it is assumed that model converged}
\item{\emph{seminr_model}}{determined by checking whether reported number of
iterations of model estimation is no greater than
the maximal number of estimations (both are
reported explicitly in the object)}
\item{\emph{cSEMResults}}{taken from information included explicitly in the
object}
}
\item{niter}{the number of iterations spent to estimated the model (not
reported for \emph{miive}, because its estimation procedure is not
iterative)}
\item{warnings}{warning messages, if any (these are not captured through
the estimation, but retrieved from the object storing the
estimated model - for \emph{miive} the only possible warning is
that about negative variance estimates, for \emph{seminr} no
warnings are reported)}
}
}
\description{
Gets summaries from estimated models.
}
