% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_something.R
\name{get_mplus_niter}
\alias{get_mplus_niter}
\title{Auxiliary functions}
\usage{
get_mplus_niter(x)
}
\arguments{
\item{x}{\code{mplusObject$results$output}}
}
\value{
a scalar (time in seconds)
}
\description{
Extracts number of iterations from mplus output.
}
\details{
Works only for models for which TECH8 reporting was turned on.
}
