% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_something.R
\name{get_mplus_time}
\alias{get_mplus_time}
\title{Auxiliary functions}
\usage{
get_mplus_time(x)
}
\arguments{
\item{x}{\code{mplusObject$results$output}}
}
\value{
a scalar (time in seconds)
}
\description{
Extracts estimation time from mplus output.
}
