% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_models.R
\name{prepare_lav_syntax}
\alias{prepare_lav_syntax}
\title{Model estimation}
\usage{
prepare_lav_syntax(mapping, interactionsMapping)
}
\arguments{
\item{mapping}{a matrix describing which observed indicators are loaded by
which latent variables, typically the \code{mapping} attribute of an object
returned by the \link{create_interaction_indicators}}

\item{interactionsMapping}{a matrix describing which exogenous latent
variables construct interaction terms, typically the \code{mapping} attribute
of an object returned by the \code{generate_latent}}
}
\value{
a string with model syntax
}
\description{
Prepares \emph{lavaan} model syntax for the \emph{Kronecker/Tensor Product Method}
approach to latent interactions model specification.
}
