% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_models.R
\name{prepare_models}
\alias{prepare_models}
\title{Model estimation}
\usage{
prepare_models(observed, observedMeanC, observedResidC, interactionsMapping)
}
\arguments{
\item{observed}{a matrix with values of observed indicators (excluding
product indicators for interaction terms)}

\item{observedMeanC}{a matrix with values of observed indicators including
product indicators for interaction terms that were created using \emph{grand mean
centering} approach, typically returned by \link{create_interaction_indicators}
called with \code{center = "mean"}}

\item{observedResidC}{a matrix with values of observed indicators including
product indicators for interaction terms that were created using \emph{residual
centering} approach, typically returned by \link{create_interaction_indicators}
called with \code{center = "residual"}}

\item{interactionsMapping}{a matrix describing which exogenous latent
variables construct interaction terms, typically the \code{mapping} attribute
of an object returned by the \code{generate_latent}}
}
\value{
a list of lists: in each element the first element is a model
estimating function and other elements are arguments that should be passed
to it
}
\description{
Prepares a list of model estimating functions along with their arguments.
}
\details{
At the moment function estimates models:
\describe{
\item{unconstrMC.ML}{Product indicators, grand mean centered - unconstrained approach estimated by ML using \emph{lavaan}}
\item{unconstrMC.TSLS}{Product indicators, grand mean centered - unconstrained approach estimated by IVSEM using \emph{MIIVsem}}
\item{unconstrRC.ML}{Product indicators, residual centered - unconstrained approach estimated by ML using \emph{lavaan}}
\item{unconstrRC.TSLS}{Product indicators, residual centered - unconstrained approach estimated by IVSEM using \emph{MIIVsem}}
\item{unconstrRC.SAM}{Product indicators, residual centered - unconstrained approach estimated by Structural After Measurement (local) using \emph{lavaan}}
\item{LMS.ML}{Latent Moderated Structural implemented in \emph{Mplus} - ML (EM) estimation}
\item{LMS.Bayes}{Latent Moderated Structural implemented in \emph{Mplus} - Bayesian estimation}
\item{PLS}{Partial Least Squares-SEM implemented via \emph{seminr} package}
}
}
