% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_models.R
\name{prepare_pls_measurement_syntax}
\alias{prepare_pls_measurement_syntax}
\title{Model estimation}
\usage{
prepare_pls_measurement_syntax(mapping, interactionsMapping = NULL)
}
\arguments{
\item{mapping}{a matrix describing which observed indicators are loaded by
which latent variables, typically the \code{mapping} attribute of an object
returned by the \link{create_interaction_indicators}}

\item{interactionsMapping}{a matrix describing which exogenous latent
variables construct interaction terms, typically the \code{mapping} attribute
of an object returned by the \code{generate_latent}}
}
\value{
an object of classes \emph{measurement_model} and \emph{seminr_model} returned
by \link[seminr:constructs]{seminr::constructs}
}
\description{
Prepares \emph{seminr} description of the measurement part of a PLS model.
}
