% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Simulation flow}
\usage{
run_simulation(conditions, nIterPerCond, suffix = "")
}
\arguments{
\item{conditions}{a data frame with simulation conditions, typically
constructed using \link{prepare_conditions}}

\item{nIterPerCond}{a positive integer - number of iterations to be run for
each condition}

\item{suffix}{optionally a string - suffix that will be added to the name
of a file storing simulation results (that will be saved to the disk)}
}
\value{
(invisibly) a list three of data frames:
\describe{
\item{modelSummaries}{basic model summary statistics - see \link{get_model_summary}}
\item{structPars}{estimates of the structural part parameters -
see \link{get_model_pars}, the additional column \code{gen} stores
the parameter value used in the data generating model}
\item{measurePars}{estimates of the measurement part parameters -
see \link{get_model_pars}, the additional columns \code{gen} and
\code{genLambda} store values of the parameters used in the
data generating model}
}
Moreover, after completing each simulation condition-iteration function will
save the data frames listed above to the file
"latentInteractions_results[suffix].RData".
}
\description{
Runs the simulation.
}
\details{
For description of estimated models see \link{prepare_models}. For references to
additional details see \link{run_iteration}.
}
\examples{
\dontrun{
str(conditions)
set.seed(12345)
run_simulation(conditions[1:2, ], nIterPerCond = 1L)
}
}
