% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_something.R
\name{standardize_pars}
\alias{standardize_pars}
\title{Auxiliary functions}
\usage{
standardize_pars(structural, covs)
}
\arguments{
\item{structural}{a data frame with structural parameters, containing
columns: \code{DV}, \code{IV}, \code{est}, \code{se} and \code{std.all}}

\item{covs}{a data frame with covariances of latent variables and error
terms, containing columns: \code{lhs}, \code{rhs} and \code{est}}
}
\value{
a data frame provided with the \code{structural} parameter with
recomputed vaues of the \code{std.all} column and new column \code{std.all.se}
}
\description{
Standardizes structural parameters of the interaction terms with respect
to standard deviations of their first-order latent variables. If no
standardized parameters are provided in the \code{structural} argument, it
standardizes all structural parameters.
}
\seealso{
\code{\link[=get_model_pars]{get_model_pars()}}
}
