% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_iteration.R
\name{run_iteration}
\alias{run_iteration}
\title{Simulation flow}
\usage{
run_iteration(condition, modelSpecsOnly = FALSE)
}
\arguments{
\item{condition}{a one-row data frame with additional parameters specifying
data generation process that will be passed to data generating functions
(\link{generate_latent}, \link{generate_loadings} and \link{generate_observed})}

\item{modelSpecsOnly}{optionally a logical flag indicating that only list
of model specifications should be returned, without actually estimating the
models}
}
\value{
By default or with \code{modelSpecsOnly = FALSE} - a list of three data
frames:
\describe{
\item{modelSummaries}{basic model summary statistics - see \link{get_model_summary}}
\item{structPars}{estimates of the structural part parameters -
see \link{get_model_pars}, the additional column \code{gen} stores
the parameter value used in the data generating model}
\item{measurePars}{estimates of the measurement part parameters -
see \link{get_model_pars}, the additional columns \code{gen} and
\code{genLambda} store values of the parameters used in the
data generating model}
}
With \code{modelSpecsOnly = TRUE} a list returned by \link{prepare_models}.
}
\description{
Performs a single iteration of the simulation.
}
\seealso{
\link{prepare_models}, \link{get_model_summary}, \link{get_model_pars}
}
